refactor(app): migrate models to protobuf-generated classes and clean up constants
- Replace custom model classes with protobuf-generated classes across the app
- Update HiveService to use protobuf serialization methods (writeToJson/mergeFromJsonMap)
- Remove unused enum definitions from app_constants.dart
- Update generate_proto.sh by removing outdated reference comment
- Clean up README.md by removing redundant internationalization pattern description
- Remove hardcoded JDK path from android/gradle.properties
- Add proper imports for protobuf files in relevant controllers and services
- Update task completion logic to use protobuf-based extensions
- Modify character class selection to use protobuf enum values
- Convert integer timestamps to Int64 for protobuf compatibility
- Update forge controller to handle Int64 gold values properly
- Refactor task handling methods to work with protobuf Task objects

+++ app/lib/features/home/pages/home_page.dart
@@ -284,7 +284,7 @@ class HomePage extends GetView<HomeController> {
     }
     final level = char?.level ?? 1;
     final needed = GameConstants.expForLevel(level).toDouble();
-    return ((char?.currentExp ?? 0) / needed).clamp(0.0, 1.0).toDouble();
+    return ((char?.currentExp.toInt() ?? 0) / needed).clamp(0.0, 1.0).toDouble();
   }
 
   Widget _walletChip({required bool gold}) {
